feat: apply user data to user page (#62)
