+++ src/apis/userInfo/getUserInfo.ts
@@ -4,13 +4,16 @@ import { END_POINTS } from "@/constants/api";
 
 import { createClient } from "@/utils/supabase/client";
 
+import type { GetUserInfoResponseType } from "@/types/userInfo";
+
 export const getUserInfo = async () => {
   const supabase = createClient();
 
   const user = await supabase.auth.getUser();
-  console.log(user);
 
-  const { data } = await axiosInstance.get(`${END_POINTS.USER_INFO}?user_id=${user.data.user?.id}`);
+  const { data } = await axiosInstance.get<GetUserInfoResponseType>(
+    `${END_POINTS.USER_INFO}?user_id=${user.data.user?.id}`,
+  );
 
   return data;
 };
